pref: use object component to remove sfc-import dep.
